requirements.txt
PyQt6>=6.8.0
PyQt6-Qt6>=6.8.1
PyQt6_sip>=13.9.1
PyYAML>=6.0.2
SQLAlchemy>=2.0.37
Werkzeug>=3.1.3
# pymetasploit3>=1.0.3  # Replaced with custom RPC implementation
msgpack>=1.0.8